--- Source/Krypton Components/Krypton.Toolkit/Utilities/General.cs
@@ -1,143 +1,0 @@
-﻿#region BSD License
-/*
- * 
- * Original BSD 3-Clause License (https://github.com/ComponentFactory/Krypton/blob/master/LICENSE)
- *  © Component Factory Pty Ltd, 2006 - 2016, (Version 4.5.0.0) All rights reserved.
- * 
- *  New BSD 3-Clause License (https://github.com/Krypton-Suite/Standard-Toolkit/blob/master/LICENSE)
- *  Modifications by Peter Wagner(aka Wagnerp) & Simon Coghlan(aka Smurf-IV), et al. 2017 - 2023. All rights reserved. 
- *  
- */
-#endregion
-
-namespace Krypton.Toolkit
-{
-    #region Enumerations
-
-    /// <summary>Specifies constants defining which buttons to display on a <see cref="T:KryptonMessageBox" />.</summary>
-    public enum KryptonMessageBoxButtons
-    {
-        /// <summary>
-        ///  Specifies that the message box contains an OK button.
-        /// </summary>
-        OK = MessageBoxButtons.OK,
-
-        /// <summary>
-        ///  Specifies that the message box contains OK and Cancel buttons.
-        /// </summary>
-        OKCancel = MessageBoxButtons.OKCancel,
-
-        /// <summary>
-        ///  Specifies that the message box contains Abort, Retry, and Ignore buttons.
-        /// </summary>
-        AbortRetryIgnore = MessageBoxButtons.AbortRetryIgnore,
-
-        /// <summary>
-        ///  Specifies that the message box contains Yes, No, and Cancel buttons.
-        /// </summary>
-        YesNoCancel = MessageBoxButtons.YesNoCancel,
-
-        /// <summary>
-        ///  Specifies that the message box contains Yes and No buttons.
-        /// </summary>
-        YesNo = MessageBoxButtons.YesNo,
-
-        /// <summary>
-        ///  Specifies that the message box contains Retry and Cancel buttons.
-        /// </summary>
-        RetryCancel = MessageBoxButtons.RetryCancel,
-
-        /// <summary>
-        ///  Specifies that the message box contains Cancel, Try Again, and Continue buttons.
-        /// </summary>
-#if NET60_OR_GREATER
-            CancelTryContinue = MessageBoxButtons.CancelTryContinue
-#else
-        CancelTryContinue = 0x00000006
-#endif 
-    }
-
-
-    /// <summary>Specifies constants defining the default button on a <seealso cref="KryptonMessageBox"/>.</summary>
-    public enum KryptonMessageBoxDefaultButton
-    {
-        /// <summary>The first button on the message box is the default button.</summary>
-        Button1 = 0,
-        /// <summary>The second button on the message box is the default button.</summary>
-        Button2 = 256,
-        /// <summary>The third button on the message box is the default button.</summary>
-        Button3 = 512,
-        /// <summary>Specifies that the Help button on the message box should be the default button.</summary>
-        Button4 = 768,
-        /// <summary>The accelerator button.</summary>
-        Button5 = 1024
-    }
-
-    /// <summary>Specifies a custom color preview shape for a <seealso cref="KryptonColorButton"/>.</summary>
-    public enum KryptonColorButtonCustomColorPreviewShape
-    {
-        Circle = 0,
-        Square = 1,
-        RoundedSquare = 2,
-        None = 3
-    }
-
-    /// <summary>Positions the title on a <see cref="KryptonForm"/>.</summary>
-    public enum KryptonFormTitleStyle
-    {
-        /// <summary>Positions the title to the left (Windows 95 - 7/10/11 style).</summary>
-        Classic = 0,
-        /// <summary>Positions the title to the center (Windows 8/8.1 style).</summary>
-        Modern = 1,
-        /// <summary>Positions the title, based on OS settings.</summary>
-        Inherit = 2
-    }
-
-    /// <summary>The <see cref="KryptonInputBox"/> type.</summary>
-    public enum KryptonInputBoxType
-    {
-        /// <summary>Use a date and time input style.</summary>
-        DateAndTime = 0,
-        /// <summary>The default WinForms style.</summary>
-        Default = 1
-    }
-
-    #endregion
-
-    #region Class: AssemblyUtilities
-
-    public class AssemblyUtilities
-    {
-        //https://www.meziantou.net/getting-the-date-of-build-of-a-dotnet-assembly-at-runtime.htm
-        public static DateTime GetLinkerTimestampUTC(Assembly assembly)
-        {
-            var location = assembly.Location;
-
-            return GetLinkerTimestampUTC(location);
-        }
-
-        public static DateTime GetLinkerTimestampUTC(string filePath)
-        {
-            const int PE_HEADER_OFFSET = 60;
-
-            const int LINKER_TIMESTAMP_OFFSET = 8;
-
-            var byteBuffer = new byte[2048];
-
-            using (var file = new FileStream(filePath, FileMode.Open, FileAccess.Read, FileShare.ReadWrite))
-            {
-                file.Read(byteBuffer, 0, byteBuffer.Length);
-            }
-
-            var headerPosition = BitConverter.ToInt32(byteBuffer, PE_HEADER_OFFSET);
-
-            var secondsSinceUNIXTimeStart = BitConverter.ToInt32(byteBuffer, headerPosition + LINKER_TIMESTAMP_OFFSET);
-
-            var dateTime = new DateTime(1970, 1, 1, 0, 0, 0, DateTimeKind.Utc);
-
-            return dateTime.AddSeconds(secondsSinceUNIXTimeStart);
-        }
-    }
-
-    #endregion
-}
