+++ Source/Krypton Components/Krypton.Toolkit/Utilities/IconType.cs
@@ -1,0 +1,10 @@
+﻿namespace Krypton.Toolkit;
+
+public enum IconType
+{
+    Warning = 101,
+    Help = 102,
+    Error = 103,
+    Info = 104,
+    Shield = 106
+}
