+++ Source/Krypton Components/Krypton.Toolkit/Utilities/KryptonColorButtonCustomColorPreviewShape.cs
@@ -1,0 +1,10 @@
+﻿namespace Krypton.Toolkit;
+
+/// <summary>Specifies a custom color preview shape for a <seealso cref="KryptonColorButton"/>.</summary>
+public enum KryptonColorButtonCustomColorPreviewShape
+{
+    Circle = 0,
+    Square = 1,
+    RoundedSquare = 2,
+    None = 3
+}
