+++ Source/Krypton Components/Krypton.Toolkit/Utilities/KryptonInputBoxType.cs
@@ -1,0 +1,10 @@
+﻿namespace Krypton.Toolkit;
+
+/// <summary>The <see cref="KryptonInputBox"/> type.</summary>
+public enum KryptonInputBoxType
+{
+    /// <summary>Use a date and time input style.</summary>
+    DateAndTime = 0,
+    /// <summary>The default WinForms style.</summary>
+    Default = 1
+}
