+++ Source/Krypton Components/Krypton.Toolkit/Utilities/KryptonMessageBoxDefaultButton.cs
@@ -1,0 +1,16 @@
+﻿namespace Krypton.Toolkit;
+
+/// <summary>Specifies constants defining the default button on a <seealso cref="KryptonMessageBox"/>.</summary>
+public enum KryptonMessageBoxDefaultButton
+{
+    /// <summary>The first button on the message box is the default button.</summary>
+    Button1 = 0,
+    /// <summary>The second button on the message box is the default button.</summary>
+    Button2 = 256,
+    /// <summary>The third button on the message box is the default button.</summary>
+    Button3 = 512,
+    /// <summary>Specifies that the Help button on the message box should be the default button.</summary>
+    Button4 = 768,
+    /// <summary>The accelerator button.</summary>
+    Button5 = 1024
+}
